dexter_sort_pipeline-main/dexter_sort_pipeline-main/pipeline/requirements.txt
gspread==6.1.2
google-auth==2.29.0
pandas==2.2.2
requests==2.31.0
python-dateutil==2.9.0
google-api-python-client
google-auth
